Renamed and removed empty file
